requirements.txt
pandas
numpy
openai
requests
python-dotenv